refactor(ingest): extract IngestPlan contract types into plan.py
Split the adapter-facing plan contract out of the 1500-line apply.py
engine into a new ingestion/plan.py: IngestPlan, the Planned* carriers,
CitationRef, CiteHandle, EntryIndex, the ResolveHook/PreWriteHook
Protocols and the RunReport result type. Repoint all importers
(adapters, patches package, ingest_patches command, tests) at plan.py
directly; apply.py now exports only apply_plan. Engine-internal
intermediates (RetractEntry, the per-entry provenance maps) stay in
apply.py — they are apply-time mechanics, not part of the contract.

Pure code motion, no behavior change.
