+++ backend/apps/catalog/tests/test_status.py
@@ -6,11 +6,11 @@ import pytest
 from django.core.exceptions import ValidationError
 from django.db import IntegrityError
 
-from apps.catalog.ingestion.apply import (
+from apps.catalog.ingestion.apply import apply_plan
+from apps.catalog.ingestion.plan import (
     IngestPlan,
     PlannedClaimAssert,
     PlannedEntityCreate,
-    apply_plan,
 )
 from apps.catalog.models import Manufacturer
 from apps.provenance.models import Claim, Source
